useformatNumber hook implemented
