+++ src/hooks/useFormatNumber.mdx
@@ -1,0 +1,14 @@
+import { Meta, Source } from "@storybook/blocks";
+
+<Meta title="Hooks/useFormatNumber" />
+
+# useFormatNumber
+
+| Hook Version | DS Version |
+| ------------ | ---------- |
+| Added        | `release`    |
+| Latest       | `release`    |
+
+`useFormatNumber` provides a number with commas and handles a range of numbers.
+
+Note: This hook is currently not being used.
