+++ src/index.ts
@@ -311,6 +311,7 @@ export { default as useNYPLBreakpoints } from "./hooks/useNYPLBreakpoints";
 export { default as useNYPLTheme } from "./hooks/useNYPLTheme";
 export { default as useWindowSize } from "./hooks/useWindowSize";
 export { default as useScrollFadeStyles } from "./hooks/useScrollFadeStyles";
+export { default as useFormatNumber } from "./hooks/useFormatNumber";
 export { default as VideoPlayer } from "./components/VideoPlayer/VideoPlayer";
 export type {
   VideoPlayerAspectRatios,
